03/15/2018 - Finished Text Based Library
The text based library is complete, now the GUI needs to be implemented.

The program is going to evolve the more that I use it and creating a GUI
interface where different things can be added changed and manipulated
will greatly help Kara in making this a functional program that she can
use in the classroom. 

Most of the bugs will show themselves in the future, and we will add
features after we have a functional library.