04/08/2018
- Finished adding most of Kara's books to the list. I am now working on
getting the actuall functions of the library now that I have something
to test them with.
- I also need to start my JUnit test cases to make sure, that as I keep
building this project bigger and more complicated that the base
components are not the reason for the failure of it. Need to start with
my hierachy with Book, Student, and Library and work my way through
them, but this can be done later on.